sourceData/bpv-fetal-infer/requirements.txt
tqdm~=4.66.2
scikit-image
nibabel==5.2.0
pydicom==2.4.4
dicom2nifti==2.4.9
monai==1.3.0
pillow~=9.3.0
flask~=3.0.3